Added Auto Indenter

+++ lilia/core/itemization/items/base/sh_bags.lua
@@ -1,4 +1,4 @@
-ITEM.name = "Bag"
+﻿ITEM.name = "Bag"
 ITEM.desc = "A bag to hold more items."
 ITEM.model = "models/props_c17/suitcase001a.mdl"
 ITEM.category = "Storage"
@@ -14,10 +14,7 @@ ITEM.functions.View = {
         if not inventory then return false end
         local panel = lia.gui["inv" .. inventory:getID()]
         local parent = item.invID and lia.gui["inv" .. item.invID] or nil
-        if IsValid(panel) then
-            panel:Remove()
-        end
-
+        if IsValid(panel) then panel:Remove() end
         if inventory then
             local panel = lia.inventory.show(inventory, parent)
             if IsValid(panel) then
@@ -29,7 +26,6 @@ ITEM.functions.View = {
             local index = item:getData("id", "nil")
             ErrorNoHalt("Invalid inventory " .. index .. " for bag item " .. itemID .. "\n")
         end
-
         return false
     end,
     onCanRun = function(item) return not IsValid(item.entity) and item:getInv() end
@@ -66,9 +62,7 @@ end
 
 function ITEM:onRemoved()
     local invID = self:getData("id")
-    if invID then
-        lia.inventory.deleteByID(invID)
-    end
+    if invID then lia.inventory.deleteByID(invID) end
 end
 
 function ITEM:getInv()
@@ -77,9 +71,7 @@ end
 
 function ITEM:onSync(recipient)
     local inventory = self:getInv()
-    if inventory then
-        inventory:sync(recipient)
-    end
+    if inventory then inventory:sync(recipient) end
 end
 
 function ITEM.postHooks:drop()
@@ -109,9 +101,7 @@ end
 if SERVER then
     function ITEM:onDisposed()
         local inventory = self:getInv()
-        if inventory then
-            inventory:destroy()
-        end
+        if inventory then inventory:destroy() end
     end
 
     function ITEM:resolveInvAwaiters(inventory)
@@ -133,7 +123,6 @@ if SERVER then
             self.awaitingInv = self.awaitingInv or {}
             self.awaitingInv[#self.awaitingInv + 1] = d
         end
-
         return d
     end
 end
