+++ lilia/modules/experimental_modules/networking/dependencies/sv_netmessagelogger.lua
@@ -1,10 +1,8 @@
-local logs = {}
+﻿local logs = {}
 util.AddNetworkString("net_ReceiveLogs")
 util.AddNetworkString("net_RequestLogs")
 local function Init()
-    if not file.Exists("netmessagelogsdir", "DATA") then
-        file.CreateDir("netmessagelogsdir")
-    end
+    if not file.Exists("netmessagelogsdir", "DATA") then file.CreateDir("netmessagelogsdir") end
 end
 
 Init()
@@ -42,10 +40,7 @@ function net.Incoming(len, ply)
     if #logs >= 5000 then
         local files = file.Find("netmessagelogsdir/*.txt", "DATA", "dateasc")
         local count = #files + 1
-        if count > 50 then
-            file.Delete("netmessagelogsdir/" .. files[1])
-        end
-
+        if count > 50 then file.Delete("netmessagelogsdir/" .. files[1]) end
         file.Write("netmessagelogsdir/" .. os.date("%d-%m-%Y_%H-%M-%S", os.time()) .. "_netmessagelogs.txt", util.TableToJSON(logs))
         logs = {}
     end
